+++ src/app-store/state.js
@@ -5,6 +5,7 @@ import isDevMode from "../utils/isDevMode";
 import masterPortalVersionNumber from "../utils/masterPortalVersionNumber";
 
 const state = {
+    _store: null,
     configJson: null,
     configJs: null,
     mobile: isMobile(), // resize update in ./index.js