+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Changed
 - The language of the changelog has been changed to English.
+- The measure tool has been migrated to Vue.
 - Migrated the Download Tool from Backbone.js to Vue.js and integrated it into the Draw Tool.
 - Migrated the Contact Tool from Backbone.js to Vue.js. In doing so, the functionality of the config parameter 'deleteAfterSend' was split into two parameters; 'closeAfterSend' and 'deleteAfterSend'.